refactor(package): 불필요하게 domain 내에 상세히 분류된 패키지를 domain으로 빼냄
