+++ src/main/kotlin/racingcar/domain/CarRandomMovementRule.kt
@@ -1,4 +1,4 @@
-package racingcar.domain.rule
+package racingcar.domain
 
 import kotlin.random.Random
