+++ src/main/kotlin/racingcar/domain/RacingResult.kt
@@ -1,0 +1,3 @@
+package racingcar.domain
+
+class RacingResult(val cars: List<Car>)
