+++ src/test/kotlin/racingcar/domain/rule/CarRandomMovementRuleTest.kt
@@ -2,7 +2,8 @@ package racingcar.domain.rule
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import racingcar.domain.car.Car
+import racingcar.domain.Car
+import racingcar.domain.CarMovementRule
 
 class CarRandomMovementRuleTest {
     @Test
