+++ src/test/kotlin/racingcar/domain/rule/MockUpRule.kt
@@ -1,5 +1,7 @@
 package racingcar.domain.rule
 
+import racingcar.domain.CarMovementRule
+
 class MockUpRule : CarMovementRule {
     var hasBeenCalled = false
     override val rule = {
